Update zerocopy dependency and refactor related code
- Updated the `zerocopy` dependency from version `0.7.35` to `0.8.40` in `Cargo.toml` and `Cargo.lock`, removing specific versioning for better flexibility.
- Refactored code across multiple modules to replace `FromZeroes` and `AsBytes` derives with `IntoBytes`, ensuring compatibility with the updated `zerocopy` version.
- Adjusted error handling and data structures in the IPC and ACPI modules to align with the new `zerocopy` API.
- Marked the `relibc` subproject as dirty to indicate local modifications.

+++ workspace/components/strate-web-admin/src/net.rs
@@ -2,7 +2,7 @@ use alloc::string::String;
 use alloc::vec::Vec;
 use strat9_syscall::{call, data::TimeSpec, flag, number};
 
-const EAGAIN: i32 = 11;
+const EAGAIN: usize = 11;
 const MAX_EAGAIN_RETRIES: usize = 50;
 
 pub fn sleep_ms(ms: u64) {
